application/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
ftfy==6.1.3
regex==2023.10.3
git+https://github.com/openai/CLIP.git